requirements.txt
Django==4.1.3
djangorestframework==3.14.0
django-filter==22.1
django-rest-swagger==2.2.0